Fix a plotly ID error
Error info: "There are multiple `plotly_chart` elements with the same auto-generated ID. When this element is created, it is assigned an internal ID based on the element type and provided parameters. Multiple elements with the same type and parameters will cause this error. To fix this error, please pass a unique `key` argument to the `plotly_chart` element."